@@ -13,14 +13,14 @@
     "vue-router": "^3.0.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
     "@vue/test-utils": "^1.0.0-beta.22",
     "autoprefixer": "^8.6.0",
-    "babel-core": "^6.26.0",
     "babel-eslint": "^8.2.5",
-    "babel-loader": "^7.1.2",
+    "babel-loader": "^8.0.0",
     "babel-plugin-istanbul": "^4.1.6",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-preset-env": "^1.6.0",
     "chai": "^4.1.2",
     "clean-webpack-plugin": "^0.1.19",
     "copy-webpack-plugin": "^4.5.1",
@@ -33,10 +33,9 @@
     "eslint-plugin-html": "^4.0.2",
     "eslint-plugin-import": "^2.9.0",
     "eslint-plugin-vue": "^4.3.0",
-    "extract-text-webpack-plugin": "^3.0.2",
     "html-webpack-exclude-assets-plugin": "0.0.7",
     "html-webpack-plugin": "^3.2.0",
-    "karma": "^2.0.5",
+    "karma": "^3.0.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-mocha": "^1.3.0",
     "karma-sourcemap-loader": "^0.3.7",
@@ -46,19 +45,21 @@
     "node-sass": "^4.5.3",
     "postcss-loader": "^2.1.5",
     "pug": "^2.0.3",
-    "pug-loader": "^2.4.0",
+    "pug-plain-loader": "^1.0.0",
+    "raw-loader": "^0.5.1",
     "sass-loader": "^6.0.6",
-    "stylelint": "^9.3.0",
+    "stylelint": "^9.2.0",
     "stylelint-config-standard": "^18.2.0",
     "stylelint-order": "^0.8.1",
     "stylelint-processor-html": "^1.0.0",
     "stylelint-webpack-plugin": "^0.10.5",
     "url-loader": "^1.0.1",
-    "vue-loader": "^13.0.5",
+    "vue-loader": "^15.4.1",
     "vue-pug-lint-loader": "0.0.1",
     "vue-template-compiler": "^2.4.4",
-    "webpack": "^3.6.0",
-    "webpack-dev-server": "^2.9.1",
+    "webpack": "^4.17.2",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-server": "^3.1.8",
     "webpack-extraneous-file-cleanup-plugin": "^1.0.5"
   }
 }
